+++ Versions/Temporary/Engine/Sources/3Dmotor/GShaderFX.hpp
@@ -1,5 +1,8 @@
 #pragma once
+
 #include <D3DX9.h>
+
+#include <cstdint>
 
 namespace NGfx
 {
@@ -20,14 +23,14 @@ public:
 
 	STDMETHOD(SetTransform)( D3DTRANSFORMSTATETYPE State, CONST D3DMATRIX *pMatrix );
 	STDMETHOD(SetMaterial)( CONST D3DMATERIAL9 *pMaterial );
-	STDMETHOD(SetLight)( DWORD Index, CONST D3DLIGHT9 *pLight );
-	STDMETHOD(LightEnable)( DWORD Index, BOOL Enable );
-	STDMETHOD(SetRenderState)( D3DRENDERSTATETYPE State, DWORD Value );
-	STDMETHOD(SetTexture)( DWORD Stage, LPDIRECT3DBASETEXTURE9 pTexture );
-	STDMETHOD(SetTextureStageState)( DWORD Stage, D3DTEXTURESTAGESTATETYPE Type, DWORD Value );
-	STDMETHOD(SetSamplerState)( DWORD Sampler, D3DSAMPLERSTATETYPE Type, DWORD Value );
+	STDMETHOD(SetLight)( unsigned long Index, CONST D3DLIGHT9 *pLight );
+	STDMETHOD(LightEnable)(unsigned long Index, BOOL Enable );
+	STDMETHOD(SetRenderState)( D3DRENDERSTATETYPE State, unsigned long Value );
+	STDMETHOD(SetTexture)(unsigned long Stage, LPDIRECT3DBASETEXTURE9 pTexture );
+	STDMETHOD(SetTextureStageState)(unsigned long Stage, D3DTEXTURESTAGESTATETYPE Type, unsigned long Value );
+	STDMETHOD(SetSamplerState)(unsigned long Sampler, D3DSAMPLERSTATETYPE Type, unsigned long Value );
 	STDMETHOD(SetNPatchMode)( FLOAT NumSegments );
-	STDMETHOD(SetFVF)( DWORD FVF );
+	STDMETHOD(SetFVF)(unsigned long FVF );
 	STDMETHOD(SetVertexShader)( LPDIRECT3DVERTEXSHADER9 pShader );
 	STDMETHOD(SetVertexShaderConstantF)( UINT RegisterIndex, CONST FLOAT *pConstantData, UINT RegisterCount );
 	STDMETHOD(SetVertexShaderConstantI)( UINT RegisterIndex, CONST INT *pConstantData, UINT RegisterCount );
